# swagger.json
{"openapi":"3.0.0","paths":{"/api/wallet/generate-mnemonic":{"post":{"description":"Generate encrypted BIP39 mnemonic for your application.\n\n**IMPORTANT:** Run this ONCE during initial Laravel setup.\n\nSteps:\n1. Call this endpoint to generate encrypted mnemonic\n2. Store the entire `encrypted_mnemonic` object in Laravel database\n3. Use the same mnemonic for all user address generation\n\nSecurity: Mnemonic is encrypted with AES-256-GCM using MASTER_PASSWORD.","operationId":"WalletController_generateMnemonic","parameters":[],"responses":{"200":{"description":"Encrypted mnemonic generated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateMnemonicResponseDto"}}}}},"summary":"1. Generate Master Mnemonic (One-Time Setup)","tags":["Laravel Integration"]}},"/api/wallet/get-address":{"post":{"description":"Generate TRON address for a specific user.\n\n**When to call:**\n- User requests deposit address\n- User wallet page loads\n\n**How it works:**\n1. Pass encrypted mnemonic from Laravel DB\n2. Pass incremental index (user ID or counter)\n3. Pass user_id to auto-enable deposit monitoring\n4. Same index always returns same address (deterministic)\n\n**Example Flow:**\n- User 1 → index: 0 → Address: TW6nF...\n- User 2 → index: 1 → Address: TRA6K...\n\n**Auto-Monitoring:** When user_id is provided, address is automatically registered for deposit detection.","operationId":"WalletController_getAddressForUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAddressRequestDto"}}}},"responses":{"200":{"description":"Address generated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAddressResponseDto"}}}},"400":{"description":"Invalid encrypted mnemonic or decryption failed","content":{"application/json":{"schema":{"example":{"success":false,"error":"Unsupported state or unable to authenticate data"}}}}}},"summary":"2. Generate User Deposit Address (Main Integration)","tags":["Laravel Integration"]}},"/api/wallet/validate-mnemonic":{"post":{"description":"Validate BIP39 mnemonic phrase.\n\n**Use cases:**\n- Validate user-imported mnemonics\n- Check backup phrase validity\n\n**Note:** Not needed for normal operations as mnemonic is auto-generated.","operationId":"WalletController_validateMnemonic","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateMnemonicRequestDto"}}}},"responses":{"200":{"description":"Mnemonic validation result","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateMnemonicResponseDto"}}}}},"summary":"4. Validate Mnemonic (Optional Helper)","tags":["Laravel Integration"]}},"/api/wallet/validate-address":{"post":{"description":"Validate TRON address format before processing.\n\n**Use cases:**\n- Validate user withdrawal addresses\n- Check address format in forms\n- Prevent invalid address submissions\n\n**Validation checks:**\n- Starts with \"T\"\n- Base58 format\n- Valid checksum\n- Length: 34 characters","operationId":"WalletController_validateAddress","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateAddressRequestDto"}}}},"responses":{"200":{"description":"Address validation result","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateAddressResponseDto"}}}}},"summary":"3. Validate TRON Address (Optional Helper)","tags":["Laravel Integration"]}}},"info":{"title":"TRX Wallet Service - Laravel Integration API","description":"🔗 **Laravel Integration Guide**\n\nThis API provides TRON (TRX/TRC20 USDT) wallet functionality for your Laravel application.\n\n**Quick Start:**\n1. Call `/wallet/generate-mnemonic` **ONCE** during setup → Store encrypted mnemonic in Laravel DB\n2. Call `/wallet/get-address` with user_id whenever user needs deposit address\n3. Implement webhook endpoint `POST /api/webhooks/deposit` in Laravel to receive deposits\n\n**Deposit Detection:**\n- Automatic monitoring every 5 minutes\n- Webhooks sent to Laravel with HMAC signature\n- Duplicate prevention with database tracking\n\n**Security:**\n- Mnemonic encrypted with AES-256-GCM\n- Webhooks signed with HMAC-SHA256\n- Never exposes private keys via API","version":"1.0","contact":{}},"tags":[{"name":"Laravel Integration","description":"Endpoints for Laravel wallet integration"}],"servers":[],"components":{"schemas":{"EncryptedDataDto":{"type":"object","properties":{"encrypted":{"type":"string","description":"AES-256-GCM encrypted data","example":"f9396b05abeb1037fc6aa3ab017ae10dadd94900f805a005a5ff362cf4e0b4cf..."},"iv":{"type":"string","description":"Initialization vector for decryption","example":"b88c7c185a2012327dc11caec3d7e9fd"},"salt":{"type":"string","description":"Salt used for key derivation","example":"2bfd374a44fa761eb54f9964dbc12778ce5e757e50f5fbd9efe25b8183d854bf"},"authTag":{"type":"string","description":"Authentication tag for data integrity","example":"372341bb8cd9b0667937fbe68650be07"}},"required":["encrypted","iv","salt","authTag"]},"GenerateMnemonicResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"encrypted_mnemonic":{"$ref":"#/components/schemas/EncryptedDataDto"},"message":{"type":"string","example":"Mnemonic generated and encrypted. Store this securely in Laravel database."}},"required":["success","encrypted_mnemonic","message"]},"GetAddressRequestDto":{"type":"object","properties":{"encrypted_mnemonic":{"description":"Encrypted mnemonic from Laravel database","allOf":[{"$ref":"#/components/schemas/EncryptedDataDto"}]},"index":{"type":"number","description":"Derivation index (incremental per user)","example":0,"minimum":0},"user_id":{"type":"number","description":"Laravel user ID (required for auto-monitoring)","example":1}},"required":["encrypted_mnemonic","index","user_id"]},"GetAddressResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"address":{"type":"string","description":"Generated TRON address","example":"TW6nF3VcaNgjWsxCHJ6F1PqHiynSxfP5KK"},"index":{"type":"number","description":"Derivation index used","example":0},"derivation_path":{"type":"string","description":"BIP44 derivation path","example":"m/44'/195'/0'/0/0"},"monitoring":{"type":"boolean","description":"Whether address is being monitored for deposits","example":true}},"required":["success","address","index","derivation_path","monitoring"]},"ValidateMnemonicRequestDto":{"type":"object","properties":{"mnemonic":{"type":"string","description":"12 or 24 word BIP39 mnemonic phrase","example":"abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about"}},"required":["mnemonic"]},"ValidateMnemonicResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"valid":{"type":"boolean","example":true}},"required":["success","valid"]},"ValidateAddressRequestDto":{"type":"object","properties":{"address":{"type":"string","description":"TRON address to validate","example":"TW6nF3VcaNgjWsxCHJ6F1PqHiynSxfP5KK"}},"required":["address"]},"ValidateAddressResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"valid":{"type":"boolean","example":true},"address":{"type":"string","example":"TW6nF3VcaNgjWsxCHJ6F1PqHiynSxfP5KK"}},"required":["success","valid","address"]}}}}